Update MovementUpdated.cs
This movement code also contains the character flip and the wall bounce feature.
the reason I created the character flip the way I did is because this way reduces lag, in stead of every frame it trying to flip the character to where it is, it checks as to whether the character is in the position and if it isn't then it flips.
The wall bounce is still very bear bones, even though it works im going to try and help make it run a little smoother.
The code makes it so it has a charging jump and the longer you hold the larger the jump and the further distance you travel. You always jump in the direction you are facing. I've also made some basic animations just to learn the ropes. I still need to add a jump animation as well as a falling animation.